split the main handle function
Split the function to ease the reading and to prepare batch handling in the
future.

This commit also changes the behaviour of the id member in case of an
"invalid request" error. Before this change the parser tried as hard as it
can to unpack the id member to use it in the response. The JSON
specification is not really clear in this case. After this change, every
"parse error" and "invalid request" error will set the id to null.